adding networkLayer
